+++ frontend/src/lib/components/FileUpload.svelte
